+++ docs/conf.py
@@ -1,6 +1,33 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
+
+
+#####################################
+#
+#   This section was suggested by ReadTheDocs here:
+#   http://read-the-docs.readthedocs.io/en/latest/faq.html#i-get-import-errors-on-libraries-that-depend-on-c-modules
+#
+#    Apparently, they try to do a complete build of your project, but they
+#    choke when they try to import and build numpy and pandas, due to these
+#    relying extensively on C modules. So I need to mock these out.
+#
+####################Added Stuff Below ################################
+
+
+import sys
+from unittest.mock import MagicMock
+
+class Mock(MagicMock):
+    @classmethod
+    def __getattr__(cls, name):
+            return Mock()
+
+MOCK_MODULES = ['pygtk', 'gtk', 'gobject', 'argparse', 'numpy', 'pandas']
+sys.modules.update((mod_name, Mock()) for mod_name in MOCK_MODULES)
+
+################ Added Stuff Above ###################################
+
 # hydrofunctions documentation build configuration file, created by
 # sphinx-quickstart on Tue Jul  9 22:26:36 2013.
 #
@@ -111,8 +138,20 @@ pygments_style = 'sphinx'
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = 'default'
+# The new default is alabaster.
+# The old default is now 'classic'.
+# more themes: http://www.sphinx-doc.org/en/stable/theming.html
+html_theme = 'alabaster'
 
+html_theme_options = {
+    'github_user': 'mroberge',
+    'github_repo': 'hydrofunctions',
+    'github_button': True,
+    'github_banner': True,
+    'travis_button': True,
+    'fixed_sidebar': True,
+    'analytics_id': 'UA-73178522-4'
+}
 # Theme options are theme-specific and customize the look and feel of a
 # theme further.  For a list of options available for each theme, see the
 # documentation.
@@ -136,7 +175,7 @@ html_theme = 'default'
 # The name of an image file (within the static path) to use as favicon
 # of the docs.  This file should be a Windows icon file (.ico) being
 # 16x16 or 32x32 pixels large.
-#html_favicon = None
+html_favicon = 'favicon.ico'
 
 # Add any paths that contain custom static files (such as style sheets)
 # here, relative to this directory. They are copied after the builtin
@@ -154,6 +193,14 @@ html_static_path = ['_static']
 
 # Custom sidebar templates, maps document names to template names.
 #html_sidebars = {}
+
+# This will add links to all of the documents in the sidebar.
+html_sidebars = {'**': [
+    'globaltoc.html',
+    'relations.html',
+    'sourcelink.html',
+    'searchbox.html',
+    ],}
 
 # Additional templates that should be rendered to pages, maps page names
 # to template names.
@@ -273,3 +320,4 @@ texinfo_documents = [
 
 # If true, do not generate a @detailmenu in the "Top" node's menu.
 #texinfo_no_detailmenu = False
+