+++ hydrofunctions/hydrofunctions.py
@@ -1,20 +1,20 @@
 # -*- coding: utf-8 -*-
 """
+hydrofunctions.py
 
+This module contains the main functions used in an interactive session.
 """
 from __future__ import absolute_import, print_function
 import requests
 import numpy as np
 import pandas as pd
-from hydrofunctions import exceptions
-
-
-def raiseit():
-    raise exceptions.HydroNoDataError
+# Change to relative import: from . import exceptions
+# https://axialcorps.com/2013/08/29/5-simple-rules-for-building-great-python-packages/
+from . import exceptions
 
 
 def get_nwis(site, service, start_date, end_date):
-    """request stream gauge data from the USGS NWIS.
+    """Request stream gauge data from the USGS NWIS.
 
     Args:
         site (str):
@@ -28,13 +28,27 @@ def get_nwis(site, service, start_date, end_date):
 
     Returns:
         a response object.
-            response.url: the url we requested data from.
-            response.status_code:
-            response.json: the content translated as json
-            response.ok: "True" when we get a '200'
+
+            * response.url: the url we requested data from.
+            * response.status_code:
+            * response.json: the content translated as json
+            * response.ok: "True" when we get a '200'
 
     Raises:
-        ConnectionError  due to connection problems like refused connection or DNS
+        ConnectionError  due to connection problems like refused connection
+        or DNS Error.
+
+    Example::
+
+        >>> from hydrofunctions import hydrofunctions as hf
+        >>> response = hf.get_nwis('01585200', 'dv', '2012-06-01', '2012-07-01')
+
+        >>> response
+        <respones [200]>
+        >>> response.ok
+        True
+        >>> response.json()
+        *JSON ensues*
 
     The specification for this service is located here:
     http://waterservices.usgs.gov/rest/IV-Service.html
@@ -66,19 +80,46 @@ def get_nwis(site, service, start_date, end_date):
 
 
 def extract_nwis_dict(response_obj):
+    """Returns a dict object from an NWIS response object.
+    """
     nwis_dict = response_obj.json()
 
     return nwis_dict
 
 
 def extract_nwis_df(response_obj):
+    """Returns a Pandas dataframe from an NWIS response object.
+
+    Returns:
+        a pandas dataframe.
+
+    Raises:
+        HydroNoDataError  when the request is valid, but NWIS has no data for
+            the parameters provided in the request.
+    """
     nwis_dict = response_obj.json()
 
     # strip header and all metadata.
     ts = nwis_dict['value']['timeSeries']
     if ts == []:
-        print('NWIS does not have data for this request')
-        return ts
+        # raise a HydroNoDataError if NWIS returns an empty set.
+        #
+        # Ideally, an empty set exception would be raised when the request
+        # is first returned, but I do it here so that the data doesn't get
+        # extracted twice.
+        # TODO: raise this exception earlier??
+        #
+        # ** Interactive sessions should have an error raised.
+        #
+        # **Automated systems should catch these errors and deal with them.
+        # In this case, if NWIS returns an empty set, then the request
+        # needs to be reconsidered. The request was valid somehow, but
+        # there is no data being collected.
+
+        # TODO: this if clause needs to be tested.
+        raise exceptions.HydroNoDataError("The NWIS reports that it does not \
+                                            have any data for this request.")
+
     data = nwis_dict['value']['timeSeries'][0]['values'][0]['value']
 
     DF = pd.DataFrame(data, columns=['dateTime', 'value'])