+++ HISTORY.rst
@@ -5,9 +5,22 @@ History
 0.1.0 (2016-08-11)
 ------------------
 
-* Automatic system for release on PyPI didn't work.
+* Project set up.
 
-0.1.1.dev1 (2016-08-11)
+0.1.1 (2016-08-11)
+------------------
+
+* Customized project boilerplate.
+
+0.1.2 (2016-08-26)
+------------------
+
+* Set up template for custom exceptions
+* Added get_nwis function
+
+0.1.3 (2016-09-09)
 -----------------------
 
-* Second attempt to get PyPI upload to work automatically.
+* Check user inputs & raise explanatory exceptions
+* Extract data from response into a dataframe.
+* Stations object for managing data.